requirements.txt.txt
streamlit
openai
PyMuPDF
python-docx
